requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aioresponses==0.7.8
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.16.4
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.57.1
anyio==4.9.0
appnope==0.1.4
apprise==1.9.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arxiv==2.2.0
asgi-lifespan==2.1.0
asttokens==3.0.0
asyncpg==0.30.0
attrs==25.3.0
Automat==25.4.16
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
biopython==1.85
blinker==1.9.0
blis==1.3.0
cachetools==6.1.0
catalogue==2.0.10
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
cleanlab==2.6.5
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.21.1
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
constantly==23.10.4
contourpy==1.3.2
coolname==2.2.0
cryptography==45.0.5
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
datasets==3.6.0
dateparser==1.2.2
ddgs==9.1.2
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docx2txt==0.9
duckduckgo_search==8.1.1
email_validator==2.2.0
emoji==2.14.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.116.1
fastapi-users==14.0.1
fastembed==0.7.1
feedparser==6.0.11
filelock==3.18.0
filetype==1.2.0
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2025.3.0
fuzzywuzzy==0.18.0
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.44
gradio_client==1.10.4
graphviz==0.21
greenlet==3.2.3
griffe==1.7.3
grpcio==1.73.1
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
html5lib==1.1
html_chunking==0.0.4
httpcore==1.0.9
httpx==0.28.1
httpx-oauth==0.16.1
httpx-sse==0.4.1
huggingface-hub==0.33.2
humanfriendly==10.0
humanize==4.12.3
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
incremental==24.7.2
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
kombu==5.5.4
langchain==0.3.26
langchain-anthropic==0.3.17
langchain-community==0.3.27
langchain-core==0.3.68
langchain-deepseek==0.1.3
langchain-experimental==0.3.4
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
langcodes==3.5.0
langdetect==1.0.9
langgraph==0.5.1
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.72
langsmith==0.4.4
language_data==1.3.0
Levenshtein==0.27.1
lightgbm==4.6.0
loguru==0.7.3
lxml==6.0.0
makefun==1.16.0
Mako==1.3.10
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow_dataclass==8.7.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
memory-profiler==0.61.0
mmh3==5.1.0
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
murmurhash==1.0.13
mutagen==1.47.0
mypy_extensions==1.1.0
narwhals==1.45.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.2.6
oauthlib==3.3.1
olefile==0.47
onnxruntime==1.22.1
openai==1.93.0
opencv-python==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.35.0
orjson==3.10.18
ormsgpack==1.10.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.3.0
paramiko==3.5.1
parsel==1.10.0
parso==0.8.4
passlib==1.7.4
pathlib==1.0.1
pathlib2==2.3.7.post1
pathspec==0.12.1
patsy==1.0.1
pendulum==3.1.0
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
playwright==1.53.0
plotly==6.2.0
pluggy==1.6.0
polars==1.31.0
portalocker==2.10.1
prefect==3.4.8
preshed==3.0.10
primp==0.15.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
Protego==0.5.0
protobuf==6.31.1
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pwdlib==0.2.1
py_rust_stemmers==0.1.5
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
PyDispatcher==2.0.7
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.13.2
PyMuPDF==1.26.3
PyNaCl==1.5.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
pypdf==5.7.0
PyPDF2==3.0.1
PySocks==1.7.1
pytesseract==0.3.13
pytest==8.4.1
pytest-asyncio==1.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-iso639==2025.2.18
python-Levenshtein==0.27.1
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
python-slugify==8.0.4
python-socks==2.7.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
qdrant-client==1.14.3
queuelib==1.8.0
RapidFuzz==3.13.0
readchar==4.2.1
redis==6.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==14.0.0
rizaio==0.12.0
rpds-py==0.26.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
schedule==1.2.2
scikit-learn==1.7.0
scipy==1.16.0
Scrapy==2.13.3
seaborn==0.13.2
selenium==4.34.2
semver==3.0.4
service-identity==24.2.0
setuptools==80.9.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.14.3
SQLAlchemy==2.0.41
srsly==2.5.1
stack-data==0.6.3
starlette==0.47.1
statsmodels==0.14.5
streamlit==1.46.1
sympy==1.14.0
tenacity==9.1.2
termcolor==3.1.0
text-unidecode==1.3
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
tldextract==5.3.0
tokenizers==0.21.2
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.2
trio==0.30.0
trio-websocket==0.12.2
Twisted==25.5.0
typeguard==4.4.4
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
unstructured==0.18.2
unstructured-client==0.38.1
urllib3==2.5.0
uv==0.7.20
uvicorn==0.35.0
vine==5.1.0
w3lib==2.3.1
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wikipedia==1.4.0
wrapt==1.17.2
wsproto==1.2.0
xgboost==3.0.2
xlrd==2.0.2
xlsxwriter==3.2.5
xxhash==3.5.0
yarl==1.20.1
youtube-search==2.1.2
yt-dlp==2025.6.30
zipp==3.23.0
zope.interface==7.2
zstandard==0.23.0

youtube_transcript_api
word2number